+++ packages/base-modal/src/Component.tsx
@@ -1,6 +1,6 @@
 /* eslint-disable jsx-a11y/no-noninteractive-element-interactions */
 import React, {
-    FC,
+    ComponentType,
     forwardRef,
     KeyboardEvent,
     MouseEvent,
@@ -22,10 +22,11 @@ import cn from 'classnames';
 
 import { Backdrop as DefaultBackdrop, BackdropProps } from '@alfalab/core-components-backdrop';
 import { Portal, PortalProps } from '@alfalab/core-components-portal';
-import { browser } from '@alfalab/core-components-shared';
+import { browser, os } from '@alfalab/core-components-shared';
 import { Stack } from '@alfalab/core-components-stack';
 import { stackingOrder } from '@alfalab/stack-context';
 
+import { lockScroll, syncHeight, unlockScroll } from './helpers/lockScroll';
 import {
     handleContainer,
     hasScrollbar,
@@ -48,7 +49,7 @@ export type BaseModalProps = {
     /**
      * Компонент бэкдропа
      */
-    Backdrop?: FC<BackdropProps>;
+    Backdrop?: ComponentType<BackdropProps>;
 
     /**
      * Свойства для Бэкдропа
@@ -202,6 +203,11 @@ export type BaseModalProps = {
      * Реф, который должен быть установлен компонентной области
      */
     componentRef?: MutableRefObject<HTMLDivElement | null>;
+
+    /**
+     * Блокирует скролл когда модальное окно открыто. Работает только на iOS.
+     */
+    iOSLock?: boolean;
 };
 
 export type BaseModalContext = {
@@ -269,6 +275,7 @@ export const BaseModal = forwardRef<HTMLDivElement, BaseModalProps>(
             zIndex = stackingOrder.MODAL,
             componentRef = null,
             usePortal = true,
+            iOSLock = false,
         },
         ref,
     ) => {
@@ -349,6 +356,10 @@ export const BaseModal = forwardRef<HTMLDivElement, BaseModalProps>(
 
         const handleClose = useCallback<Required<BaseModalProps>['onClose']>(
             (event, reason) => {
+                if (iOSLock && os.isIOS()) {
+                    unlockScroll();
+                }
+
                 if (onClose) {
                     onClose(event, reason);
                 }
@@ -363,7 +374,7 @@ export const BaseModal = forwardRef<HTMLDivElement, BaseModalProps>(
 
                 return null;
             },
-            [onBackdropClick, onClose, onEscapeKeyDown],
+            [onBackdropClick, onClose, onEscapeKeyDown, iOSLock],
         );
 
         const handleBackdropMouseDown = (event: MouseEvent<HTMLElement>) => {
@@ -469,7 +480,13 @@ export const BaseModal = forwardRef<HTMLDivElement, BaseModalProps>(
                 if (!disableBlockingScroll) {
                     const el = getContainer();
 
-                    handleContainer(el);
+                    const shouldIOSLock = iOSLock && os.isIOS();
+
+                    handleContainer(el, shouldIOSLock);
+                    if (shouldIOSLock) {
+                        syncHeight();
+                        lockScroll();
+                    }
 
                     restoreContainerStylesRef.current = () => {
                         restoreContainerStylesRef.current = null;
@@ -479,7 +496,7 @@ export const BaseModal = forwardRef<HTMLDivElement, BaseModalProps>(
 
                 setExited(false);
             }
-        }, [getContainer, open, disableBlockingScroll, isExited]);
+        }, [getContainer, open, disableBlockingScroll, isExited, iOSLock]);
 
         useEffect(() => {
             const ResizeObserver = window.ResizeObserver || ResizeObserverPolyfill;
@@ -496,6 +513,12 @@ export const BaseModal = forwardRef<HTMLDivElement, BaseModalProps>(
                 }
             };
         }, []);
+
+        useEffect(() => {
+            if (disableAutoFocus || !open) return;
+
+            wrapperRef.current?.focus();
+        }, [open, disableAutoFocus]);
 
         const contextValue = useMemo<BaseModalContext>(
             () => ({
@@ -531,7 +554,6 @@ export const BaseModal = forwardRef<HTMLDivElement, BaseModalProps>(
                 {(computedZIndex) => (
                     <BaseModalContext.Provider value={contextValue}>
                         <FocusLock
-                            autoFocus={!disableAutoFocus}
                             disabled={disableFocusLock || !open}
                             returnFocus={!disableRestoreFocus}
                         >
@@ -564,8 +586,7 @@ export const BaseModal = forwardRef<HTMLDivElement, BaseModalProps>(
                                 onKeyDown={handleKeyDown}
                                 onMouseDown={handleBackdropMouseDown}
                                 onMouseUp={handleBackdropMouseUp}
-                                // eslint-disable-next-line jsx-a11y/no-noninteractive-tabindex
-                                tabIndex={0}
+                                tabIndex={-1}
                                 data-test-id={dataTestId}
                                 style={{
                                     zIndex: computedZIndex,