+++ packages/amount-input/CHANGELOG.md
@@ -1,5 +1,21 @@
 # Change Log
 
+## 7.9.5
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   input@14.4.3
+    -   with-suffix@4.1.27
+
+## 7.9.4
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   input@14.4.2
+    -   with-suffix@4.1.26
+
 ## 7.9.3
 
 ### Patch Changes