+++ packages/calendar-with-skeleton/CHANGELOG.md
@@ -1,5 +1,20 @@
 # Change Log
 
+## 5.2.13
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   skeleton@5.0.0
+    -   calendar@7.6.3
+
+## 5.2.12
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   calendar@7.6.2
+
 ## 5.2.11
 
 ### Patch Changes