+++ packages/checkbox/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log
 
+## 4.3.1
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   shared@0.10.0
+
 ## 4.3.0
 
 ### Minor Changes